requirements.txt
appdirs==1.4.4
arrow==1.0.2
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
distlib==0.3.1
Django==3.1.7
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-embed-video==1.4.0
django-froala-editor==3.2.2
django-js-asset==1.2.2
djangorestframework==3.12.4
filelock==3.0.12
gTTS==2.2.2
idna==2.10
numpy==1.20.1
pandas==1.2.3
Pillow==8.2.0
psutil==5.8.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
virtualenv==20.4.4
whitenoise==5.2.0
gunicorn